pyproject.toml
[project]
name = "draftly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
]
